arm64: dts: msm8996: Add MotoZ target
Add a new target for the MotoZ which is initially a copy of the msm8996 MTP
reference board. We are required to add a handful of board and pmic
revision properties in order for the device to boot on a stock
bootloader. The Model must also match the expected value for the
Motorola bootloader to find the matching DTB.

Signed-off-by: Jason Hrycay <jhrycay@gmail.com>