staging: greybus: export gb_connection_hd_find
Expose the connection find API in order for the mods_ap layer to query
for the gb_connection structure in order to help identify a CPORT's
protocol. The protocol identifier is needed in order to properly filter
or intercept certain messages from going to the end device.

This can be improved by moving some of this filtering back into the core
greybus.

Signed-off-by: Jim Wylder <jwylder@motorola.com>