staging: greybus: mods: Update protocol id retrieval
Scan through the gb_bundle's cport descriptors to find the actual
protocol ID. This previously was in the gb_connection structure but has
since been re-architected. The protocol ID is necessary in order to
support the original implementation of 'protocol message filters' which
are used to intercept certain messages from reaching the attached mod.

Signed-off-by: Jason Hrycay <jhrycay@gmail.com>