staging: greybus: mods: check global muc data
Avoid de-referencing a null pointer when checking for ack support. If
the core muc driver fails to probe, this will result in a null pointer.
Of course, we could architect this slightly different where the muc_spi
driver is actually a child and so would never probe unless the core muc
driver succeeds.

Signed-off-by: Jason Hrycay <jhrycay@gmail.com>