staging: greybus: light: reset pointers on release
Once a gb_light has been released, and its channels and name have been
free'd, reset the pointers back to null and clear the has_flash flag.

There exists at least one path when handling GB_LIGHTS_LIGHT_CONFIG that
can lead to double-free of kernel memory. When we receive this event, we
release the existing entry and perform light_config. If light_config
fails before we re-allocate the channel and name pointers when sending
GB_LIGHTS_TYPE_GET_LIGHT_CONFIG, the light_release will be invoked again
leading to a double-free of the old channels and name pointers. Similar
issues are present with the has_flash flag causing a double-unregister
with the v4l2 sub-system.

Signed-off-by: Jason Hrycay <jhrycay@gmail.com>